Architecture. Tokens. Contract Function Calling. (#2)
- [WA-11]: Maintain ERC20 token addresses
- [WA-10]: Sign and broadcast transactions
- [WA-15]: Work on app architecture

Token balances for the current account are also shown per token, but currently no whitelisted tokens exist & also no filtering for 0 balances.

Secrets are maintained via Code Generation with Sourcery in a build phase. Locally, people can add a `env-vars.sh` file which will be used during compilation, on CI the variables are set directly. An `env-vars.sh.example` file is added which can be changed to the real keys.

Tokens are currently stored on the file system via Disk. Keys are stored in the Keychain using KeychainAccess. Tests are written with Quick & Nimble.